LayoutTests/imported/w3c/web-platform-tests/pointerevents/compat/pointerevent_mouse-pointer-updown-events-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Verifies that pointerup/down are fired correctly for corresponding mouse events


Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT pointerup/down events are fired correctly for corresponding mouse events Test timed out

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT pointerup/down events are fired correctly for corresponding mouse events Test timed out
